Handles unknown bucket errors that don't require a response
Changes the handling of unknown bucket errors to never emit `error`
events. If the error is related to a request that does not need a
response the error is logged directly in the Request Handler.

By not emitting errors we avoid breaking the stream pipe which could
would end in new requests not being passed to the RequestHandler.